OA-30: Provider association to occure after authentication
